Move definitions in Arp to Arp.ArpManager
